Lodha Hinjewadi Pune Apartments
Lodha Group has launched Lodha Hinjewadi, another housing development that embodies the current design style in the entirety of its brilliance. Homeowners can take advantage of a world of unrivaled luxury in these stunning landmark properties. One of Lodha Group's best projects is this amazing Pune project with 2, 3, and 4-bedroom apartments.

With plenty of space, natural light, and cross ventilation, the homes are beautifully designed. Sitting in the expansive balconies and windows, homeowners can take in the stunning views. 75% of the property is open space. Each residential wing of the complex has three lifts, and the living room and master bedroom both have televisions. From each apartment to the security cabin, it has an intercom facility. The apartments in the complex have standby power of 1 kW and 100% backup power. The property has a clubhouse, a meditation center, a sports facility, a swimming pool, a gymnasium, a landscape garden, a park, open space, an intercom, firefighting gear, power backup, rainwater harvesting, and a water supply that is available around the clock. It has parking for cars, an earthquake-resistant structure, a multipurpose hall, video security, a sewage treatment plant, a lift, 24/7 security, a basketball court, a jogging track, a tennis court, and other amenities.

This is one of Pune's most sought-after residential projects. Stepped seating, a skating rink, a pond, a fountain plaza, and a play area for kids are all present. There is a separate park for seniors. In addition, there is a boardwalk, a cricket pitch, a multi-court, a swimming pool, a kids' pool, and a work-from-nature area on the property. 

ore Info-
Visit- https://lodha.developerprojects.in/residential/pune/lodha-hinjewadi/